Completed issue #100
Plots all component label dists
