+++ configs/stats.ini
@@ -1,7 +1,7 @@
 [DEFAULT]
 whitelist = no
 parallel  = no
-pipe      = LabelDistribution, ComponentNodeLabelStats
+pipe      = LabelDistributionPerComponent
 
 [moves]
 csv  = /Users/justin/Documents/path/data/master/moves.tsv
